--- .editorconfig
@@ -1,14 +1,0 @@
-root = true
-
-[*]
-charset = utf-8
-end_of_line = lf
-insert_final_newline = true
-indent_style = space
-trim_trailing_whitespace = true
-
-[*.py]
-indent_size = 4
-
-[*.md]
-trim_trailing_whitespace = false
